requirements.txt
python >= 3.5
attrs